examples/shopping-cart/backend/requirements.txt
# Shopping Assistant Backend Requirements

# Core dependencies
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-dotenv>=1.0.0

# Agent State Bridge
agent-state-bridge[fastapi]>=0.2.0

# LangChain and AI
langchain-openai>=0.1.0
langchain-core>=0.2.0

# Optional: For RAG/Vector stores (future enhancement)
# langchain-community>=0.2.0
# chromadb>=0.4.0
# sentence-transformers>=2.2.0
